+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Version 2.0.24
 
 - Angular project now lets you choose the package manager you want to use
+- Package manager default selection is now a setting
 
 ### Version 2.0.23