+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Angular project now lets you choose the package manager you want to use
 - Package manager default selection is now a setting
+- Debugging for Vite based projects on Android is now working
 
 ### Version 2.0.23